requirements.txt
# AI Resume Analyzer Dependencies
# Core NLP and ML
spacy>=3.7.0
scikit-learn>=1.3.0
nltk>=3.8.0

# PDF Processing
PyMuPDF>=1.23.0
pdfminer.six>=20221105

# Data Analysis
pandas>=2.0.0

# Web Interface
streamlit>=1.28.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0
plotly>=5.17.0

# AI/LLM Integration
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
tiktoken>=0.5.0

# Development Dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
